backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-docx==1.1.0
markdown==3.5.1
anthropic==0.18.1
python-dotenv==1.0.0
redis==5.0.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
PyJWT==2.8.0
bcrypt==4.1.2
email-validator==2.1.0
gunicorn==21.2.0
stripe==7.9.0
sendgrid==6.11.0
sentry-sdk[flask]==1.39.2
